Include support for YAML files with multiple documents in stream
YAML supports multiple documents in one stream. This entails that
YAML files could be viewed as a forest rather than a single tree.
Graphtage supports only tree-like files.

Support multi-document YAML files by considering each document as an
array element, i.e. consider each document as part of the same tree.
These changes is fully backwards compatible as it considers
single-document YAML streams as a single node.

An alternative approach to this solution would be to go through each
document as separate trees and diffing them. However, this approach
would include overhead in changes, such as traversing each tree
separately and considering trees of different depth.

An addition could be to sort the array before diffing. However, this
addition would not reflect the reality of the files as the order of
documents are dropped.